requirements.txt
robotframework==7.4.2
robotframework-browser==19.12.5
robotframework-requests==0.9.7
robotframework-pabot==2.20.0

# GitHub integration (used by scripts/)
PyGithub==2.3.0
requests==2.32.5
python-dotenv==1.0.1

# DOM Inspector
beautifulsoup4==4.14.3

# OpenCV Visual Debug
opencv-python==4.13.0.92
numpy==2.4.3
